bot/requirements.txt
python-telegram-bot==20.7
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
web3>=6.0.0
python-dotenv>=1.0.0
